backend/requirements.txt
fastapi
numpy
groq
pillow
python-multipart
python-dotenv
opencv-python
opencv-contrib-python
ultralytics
uvicorn
requests